conscience: dumb master/slave replication
Specify "master_conscience" in configuration file of slave
consciences with the IP:PORT of the conscience acting as master.
The slave conscience will register on master conscience all
services that have registered locally, and fetch the list of
all services with their scores, each second.

This allows static load balancing.

Change-Id: Iea6e3917a434e4d63661ee7ee25085ccbd2f9dae