[WIP] metautils/lb: add possibility to filter service list before search
This is an attempt to enhance performance of load balancing by filtering
out services which do not match a custom criteria before the actual
search.

For example this can be used to search for services defined with a
certain tag: the services without this tag are filtered out before load
balancing algorithm is performed.

Note that this is implemented using a partial copy of the service
references, thus can impact performance negatively. There should be no
noticeable loss of performance if filter is not used (in which case no
copy is made). However this has yet to be thoroughly tested and benched
(hence the WIP tag, to be removed after benchmarks).

Change-Id: Ib6bc52c3260ff52a22535e52760f748cd2678c89