LB: build a shuffled list of services
Previous behavior was to use a tree to build up the set of services to use,
then convert this tree to an array and then shuffle the array. In addition the
shuffling algorithm was not working very well (proved by benchmarks).

This commit aims to simplify this process by directly building a shuffled array
of services.

Change-Id: I04dedc646150a3dd8fcbbe9ec91d77eb9db73f00